Cleanup structure
Moves many shared types and functions out of `languageFeatures` into `types` or `util`
