+++ src/languageFeatures/codeActions/removeLinkDefinition.ts
@@ -5,11 +5,11 @@
 
 import * as l10n from '@vscode/l10n';
 import * as lsp from 'vscode-languageserver-protocol';
+import { MdLinkDefinition } from '../../types/documentLink';
 import { rangeIntersects } from '../../types/range';
 import { getDocUri, ITextDocument } from '../../types/textDocument';
 import { WorkspaceEditBuilder } from '../../util/editBuilder';
 import { DiagnosticCode } from '../diagnostics';
-import { MdLinkDefinition } from '../documentLinks';
 import { codeActionKindContains } from './util';
