+++ src/test/documentLinks.test.ts
@@ -7,8 +7,9 @@ import * as assert from 'assert';
 import * as lsp from 'vscode-languageserver-protocol';
 import { URI } from 'vscode-uri';
 import { getLsConfiguration } from '../config';
-import { InternalHref, MdLink, MdLinkComputer, MdLinkProvider } from '../languageFeatures/documentLinks';
+import { MdLinkComputer, MdLinkProvider } from '../languageFeatures/documentLinks';
 import { MdTableOfContentsProvider } from '../tableOfContents';
+import { InternalHref, MdLink } from '../types/documentLink';
 import { InMemoryDocument } from '../types/inMemoryDocument';
 import { noopToken } from '../util/cancellation';
 import { ContainingDocumentContext, IWorkspace } from '../workspace';
